Feat: Print answer in display and behaviour change
Add ability to print answer in the calculator
display, and change the behaviour of the display.

Previously the behaviour was as follows: display
all output on single line. For example: 1 + 2 = 3

Now the first operand will be displayed on line 1
and moved above to the background on line 2, when
a operator has been selected. The second operand
will then be moved to line 2 also when the equals
is pressed. The answer will display on line 1. In
other words the bottom line (1) is the higher
priority line, and line 2 is the background line.
For example:
-------------------------------------------------
-                                               -
-                                               -
- 1                                             -
-------------------------------------------------

-------------------------------------------------
- 1 +                                           -
-                                               -
- 2                                             -
-------------------------------------------------

-------------------------------------------------
- 1 + 2                                         -
-                                               -
- = 3                                           -
-------------------------------------------------